+++ src/api/metrics.cpp
@@ -48,14 +48,14 @@ f64arr precision_recall_2d(const i64arr& conf_mat, const double fill) {
         throw std::runtime_error("`conf_mat` should have shape (N, 4)");
     }
 
-    const size_t n_obs = conf_mat.shape(0);
-    auto metrics = py::array_t<double>({n_obs, static_cast<size_t>(2)});
+    const int64_t n_obs = conf_mat.shape(0);
+    auto metrics = py::array_t<double>({n_obs, static_cast<int64_t>(2)});
 
     int64_t* const cm_ptr = npy::get_data(conf_mat);
     double* const metrics_ptr = npy::get_data(metrics);
     // compute metrics
     #pragma omp parallel for shared(cm_ptr, metrics_ptr)
-    for (size_t i = 0; i < n_obs; i++) {
+    for (int64_t i = 0; i < n_obs; i++) {
         core::precision_recall(cm_ptr + (i * 4), metrics_ptr + (i * 2), fill);
     }
     return metrics;
@@ -79,15 +79,15 @@ f64arr precision_recall_flattened(const i64arr& conf_mat, const double fill) {
         throw std::runtime_error("`conf_mat` should have shape (N * 4)");
     }
 
-    size_t n_conf_mats = conf_mat.size() / 4;
+    int64_t n_conf_mats = conf_mat.size() / 4;
     int64_t* const cm_ptr = npy::get_data(conf_mat);
 
     // allocate memory for metrics; are all set so don't rely on initialisation
-    auto metrics = py::array_t<double>(n_conf_mats * static_cast<ssize_t>(2));
+    auto metrics = py::array_t<double>(n_conf_mats * static_cast<int64_t>(2));
     double* const metrics_ptr = npy::get_data(metrics);
 
     #pragma omp parallel for shared(cm_ptr, metrics_ptr)
-    for (size_t i = 0; i < n_conf_mats; i++) {
+    for (int64_t i = 0; i < n_conf_mats; i++) {
         // compute metrics
         core::precision_recall(cm_ptr + (i * 4), metrics_ptr + (i * 2), fill);
     }
@@ -134,14 +134,14 @@ f64arr binary_metrics_2d(const i64arr& conf_mat, const double fill) {
         throw std::runtime_error("`conf_mat` should have shape (N, 4)");
     }
 
-    const size_t n_obs = conf_mat.shape(0);
+    const int64_t n_obs = conf_mat.shape(0);
     int64_t* const cm_ptr = npy::get_data(conf_mat);
-    auto metrics = py::array_t<double>({n_obs, static_cast<size_t>(10)});
+    auto metrics = py::array_t<double>({n_obs, static_cast<int64_t>(10)});
     double* const metrics_ptr = npy::get_data(metrics);
 
     // compute metrics
     #pragma omp parallel for shared(cm_ptr, metrics_ptr)
-    for (size_t i = 0; i < n_obs; i++) {
+    for (int64_t i = 0; i < n_obs; i++) {
         core::binary_metrics(cm_ptr + (i * 4), metrics_ptr + (i * 10), fill);
     }
     return metrics;
@@ -166,15 +166,15 @@ f64arr binary_metrics_flattened(const i64arr& conf_mat, const double fill) {
         throw std::runtime_error("`conf_mat` should have shape (N * 4)");
     }
 
-    size_t n_conf_mats = conf_mat.size() / 4;
+    int64_t n_conf_mats = conf_mat.size() / 4;
     int64_t* const cm_ptr = npy::get_data(conf_mat);
 
     // allocate memory for metrics; are all set so don't rely on initialisation
-    auto metrics = py::array_t<double>(n_conf_mats * static_cast<ssize_t>(10));
+    auto metrics = py::array_t<double>(n_conf_mats * static_cast<int64_t>(10));
     double* const metrics_ptr = npy::get_data(metrics);
 
     #pragma omp parallel for shared(cm_ptr, metrics_ptr)
-    for (size_t i = 0; i < n_conf_mats; i++) {
+    for (int64_t i = 0; i < n_conf_mats; i++) {
         // compute metrics
         core::binary_metrics(cm_ptr + (i * 4), metrics_ptr + (i * 10), fill);
     }
